+++ stdlib.sh
@@ -238,8 +238,10 @@ dotenv() {
   elif [[ -d $path ]]; then
     path=$path/.env
   fi
-  watch_file "$path"
-  if ! [[ -f $path ]]; then
+  if ! [[ -p $path ]]; then
+    watch_file "$path"
+  fi
+  if ! [[ -f $path || -p $path ]]; then
     log_error ".env at $path not found"
     return 1
   fi
@@ -257,8 +259,10 @@ dotenv_if_exists() {
   elif [[ -d $path ]]; then
     path=$path/.env
   fi
-  watch_file "$path"
-  if ! [[ -f $path ]]; then
+  if ! [[ -p $path ]]; then
+    watch_file "$path"
+  fi
+  if ! [[ -f $path || -p $path ]]; then
     return
   fi
   eval "$("$direnv" dotenv bash "$@")"